+++ simulation/include/imu/numerical_diff_imu.h
@@ -20,17 +20,17 @@ public:
     
     void initialize(VioSimulation *simulation);
     
-    void runLoopCallback(float time);
+    void runLoopCallback(double time);
     
-    virtual Eigen::Vector3d getAccelerometerData(float time) const;
-    virtual Eigen::Vector3d getGyroscopeData(float time) const;
+    virtual Eigen::Vector3d getAccelerometerData(double time) const;
+    virtual Eigen::Vector3d getGyroscopeData(double time) const;
     
     virtual ~NumericalDiffImu();
     
 protected:
     NumericalDiffImu(SimSetup *sim_setup);
     
-    Eigen::Vector3d getVelocity(float time) const;
+    Eigen::Vector3d getVelocity(double time) const;
     
     template <typename T>
     T diff(std::function<T(double)> f, double x, double h) const {
